Change precision from float64 to float32 - to match c++ version goes down from 1.14 to 0.95. Small inconvinience is intoduction of explicit conversions as float64 is the default type. Not sure if this change is justified.

+++ _examples/main.go
@@ -41,7 +41,7 @@ func (b B) Hello() {
 	fmt.Println(b.x)
 }
 
-func hit_sphere(center Vec3, radius float64, r Ray) bool {
+func hit_sphere(center Vec3, radius float32, r Ray) bool {
 	oc := r.Origin().Subtr(center)
 	a := r.Direction().Dot(r.Direction())
 	b := oc.Dot(r.Direction()) * 2.0
@@ -112,11 +112,11 @@ func main() {
 	_ = focal_length
 
 	origin := NewVec3(0, 0, 0)
-	horizontal := NewVec3(viewport_width, 0, 0)
-	vertical := NewVec3(0, viewport_height, 0)
+	horizontal := NewVec3(float32(viewport_width), 0, 0)
+	vertical := NewVec3(0, float32(viewport_height), 0)
 
 	lower_left_corner := origin.Subtr(horizontal.DivF(2.0))
-	lower_left_corner = lower_left_corner.Subtr(vertical.DivF(2.0)).Add(NewVec3(0, 0, focal_length))
+	lower_left_corner = lower_left_corner.Subtr(vertical.DivF(2.0)).Add(NewVec3(0, 0, float32(focal_length)))
 
 	upLeft := image.Point{0, 0}
 	lowRight := image.Point{width, height}
@@ -130,8 +130,8 @@ func main() {
 			// g = uint8(255 * float64(j) / float64(height))
 			// img.SetRGBA(i, height-j, color.RGBA{r, g, 0, 255})
 
-			u := float64(i) / float64(width-1)
-			v := float64(j) / float64(height-1)
+			u := float32(i) / float32(width-1)
+			v := float32(j) / float32(height-1)
 
 			dir := lower_left_corner.Add(horizontal.MultF(u)).Add(
 				vertical.MultF(v)).Subtr(origin)
